Fix keycloak config username key
Currently the key used to extract a users username from Keycloak
is preferred_name, this is incorrect, the correct key is
preferred_username.

This currently causes retrieving the current user to fail unless a
first name has been set on the user.

This commit updates the key from preferred_name to preferred_username.