Deprecate getRoles function due to naming
Currently there is a function User#getRoles. This does not follow
the recommended naming convention for C# and does not follow the
naming conventions for the rest of the functions in the SDKs.

There is also a public field exposed on an instance of User named
Roles. This is not readonly and can be overwritten by an end user,
it exposes the list of roles that the user contains.

This change deprecates the function User#getRoles and recommends
to instead use the property Roles. It also refactors Roles from a
field into a readonly property without changing the return type
that the end user expects.

This means, apart from Roles no longer being modifiable for an end
user (which they shouldn't be doing anyway), the change is backwards
compatible.